requirements.txt
# Core
requests>=2.31
beautifulsoup4>=4.11
lxml>=4.9
pyyaml>=6.0
pandas>=2.0
numpy>=1.24
scipy>=1.10

# Bio
anndata>=0.9
scanpy>=1.9
GEOparse>=2.0
pronto>=2.5

# Embedding / reranking
sentence-transformers>=2.7
FlagEmbedding>=1.2
qdrant-client>=1.7

# LLM serving / client
ollama>=0.3

# NLP / UX
spacy>=3.7
rich>=13.0

# Utilities
tqdm>=4.65
pydantic>=2.0
